ignore nokogiri profiling code in test coverage
